requirements.txt
streamlit
openai-whisper
ffmpeg
ffmpeg-python
python-docx
reportlab